requirements.txt
azure-ai-projects
python-dotenv
azure-identity
azure-ai-evaluation